+++ packages/scripts/src/versions/index.ts
@@ -1,0 +1,7 @@
+import { printWorkspacePackages } from './lib'
+
+async function main() {
+  printWorkspacePackages()
+}
+
+main()