+++ packages/composites/vesper/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/vesper-adapter
 
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ea-bootstrap@1.9.0
+  - @chainlink/token-allocation-adapter@1.0.29
+
 ## 1.0.28
 
 ### Patch Changes